+++ blocklets/ai-runtime/vite.config.ts
@@ -1,7 +1,5 @@
 /* eslint-disable import/no-extraneous-dependencies */
-import path from 'path';
 
-import buildOpenAPIPlugin from '@blocklet/dataset-sdk/plugin';
 import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
 import { createBlockletPlugin } from 'vite-plugin-blocklet';
@@ -13,12 +11,7 @@ export default defineConfig(() => {
     optimizeDeps: {
       force: true,
     },
-    plugins: [
-      tsconfigPaths(),
-      react(),
-      createBlockletPlugin(),
-      buildOpenAPIPlugin({ apis: [path.join(__dirname, './api/src/routes/**/*.*')] }),
-    ],
+    plugins: [tsconfigPaths(), react(), createBlockletPlugin()],
     build: {
       commonjsOptions: {
         transformMixedEsModules: true,