+++ blocklets/ai-studio/src/components/file-editor/output/AppearanceSettings.tsx
@@ -2,8 +2,6 @@ import { Component, getComponents } from '@app/libs/components';
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
 import { OutputVariableYjs, RuntimeOutputAppearance, RuntimeOutputVariable } from '@blocklet/ai-runtime/types';
 import { Map, getYjsValue } from '@blocklet/co-git/yjs';
-import { REMOTE_REACT_COMPONENT } from '@blocklet/components-sdk/const';
-import { RemoteComponent } from '@blocklet/components-sdk/type';
 import { Icon } from '@iconify-icon/react';
 import {
   Autocomplete,
@@ -17,11 +15,12 @@ import {
   Typography,
 } from '@mui/material';
 import { WritableDraft } from 'immer';
-import pick from 'lodash/pick';
-import { useEffect, useMemo } from 'react';
+import { useMemo } from 'react';
 import { useAsync } from 'react-use';
 
-import { getDynamicReactComponents } from '../../../libs/components';
+import { getOpenComponents } from '../../../libs/components';
+import { REMOTE_REACT_COMPONENT } from '../../../libs/constants';
+import { RemoteComponent } from '../../../libs/type';
 import ComponentSettings from './ComponentSettings';
 
 const ignoreAppearanceSettingsOutputs = new Set<string>([RuntimeOutputVariable.children]);
@@ -52,25 +51,6 @@ export default function AppearanceSettings({
     });
   };
 
-  // 兼容旧版本数据，2024-06-23 之后可以删掉
-  useEffect(() => {
-    if (!output.appearance && (output.initialValue as any)?.componentId) {
-      setField((appearance) => {
-        const {
-          componentId,
-          componentName,
-          componentProps: props,
-        } = (pick(output.initialValue, 'componentId', 'componentName', 'componentProps') as any) ?? {};
-
-        Object.assign(appearance, {
-          componentId,
-          componentName,
-          componentProperties: props ? JSON.parse(JSON.stringify(props)) : undefined,
-        });
-      });
-    }
-  }, []);
-
   const { tags } = useMemo(() => {
     const m: { [key: string]: { tags: string } } = {
       [RuntimeOutputVariable.appearancePage]: { tags: 'aigne-page,aigne-layout' },
@@ -83,7 +63,7 @@ export default function AppearanceSettings({
 
   const { value: remoteReact } = useAsync(
     async () =>
-      getDynamicReactComponents().then((components) =>
+      getOpenComponents().then((components) =>
         components.filter((component) => (component?.tags || []).some((tag) => tags.includes(tag)))
       ),
     [tags]
@@ -216,10 +196,7 @@ function ComponentSelect({
       ...(remoteReact || []).map((x) => ({
         id: REMOTE_REACT_COMPONENT,
         name: x.name,
-        componentProperties: {
-          componentPath: x.url,
-          blockletDid: x.did,
-        },
+        componentProperties: { componentPath: x.url, blockletDid: x.did },
         group: t('remote'),
       })),
     ];
@@ -235,8 +212,11 @@ function ComponentSelect({
       getOptionLabel={(component) =>
         component.blockletDid ? component.name || component.id : `${component.name || component.id} (Local)`
       }
-      isOptionEqualToValue={(o, v) =>
-        o.id === v.id && ((!o.blockletDid && !v.blockletDid) || o.blockletDid === v.blockletDid)
+      isOptionEqualToValue={
+        (o, v) =>
+          o.id === v.id &&
+          ((!o.blockletDid && !v.blockletDid) || o.blockletDid === v.blockletDid) &&
+          ((!o.name && !v.name) || o.name === v.name) // FIXME: 临时解决方案，等后端返回 name 字段后可以删掉
       }
       renderGroup={(params) => {
         return (