+++ blocklets/ai-studio/src/components/file-editor/router-file/index.tsx
@@ -19,7 +19,7 @@ import {
 import { Map, getYjsValue } from '@blocklet/co-git/yjs';
 import { getAllParameters } from '@blocklet/dataset-sdk/request/util';
 import { DatasetObject } from '@blocklet/dataset-sdk/types';
-import getDatasetTextByI18n from '@blocklet/dataset-sdk/util/get-dataset-i18n-text';
+import getOpenApiTextFromI18n from '@blocklet/dataset-sdk/util/get-open-api-i18n-text';
 import { Icon } from '@iconify-icon/react';
 import CheckIcon from '@iconify-icons/tabler/check';
 import ArrowFork from '@iconify-icons/tabler/corner-down-right';
@@ -54,10 +54,10 @@ import {
   createFilterOptions,
   styled,
 } from '@mui/material';
-import { cloneDeep, groupBy, pick, sortBy } from 'lodash';
+import { cloneDeep, pick, sortBy } from 'lodash';
 import { bindDialog, usePopupState } from 'material-ui-popup-state/hooks';
 import { nanoid } from 'nanoid';
-import React, { forwardRef, useCallback, useImperativeHandle, useMemo, useRef } from 'react';
+import { forwardRef, useCallback, useImperativeHandle, useMemo, useRef } from 'react';
 import { Controller, UseFormReturn, useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
 import { joinURL } from 'ufo';
@@ -108,8 +108,8 @@ export default function RouterAssistantEditor({
       id: dataset.id,
       type: dataset.type,
       name:
-        getDatasetTextByI18n(dataset, 'summary', locale) ||
-        getDatasetTextByI18n(dataset, 'description', locale) ||
+        getOpenApiTextFromI18n(dataset, 'summary', locale) ||
+        getOpenApiTextFromI18n(dataset, 'description', locale) ||
         t('unnamed'),
       from: dataset.from,
     })) as RouteOption[];
@@ -545,8 +545,8 @@ export const ToolDialog = forwardRef<
       id: dataset.id,
       type: dataset.type,
       name:
-        getDatasetTextByI18n(dataset, 'summary', locale) ||
-        getDatasetTextByI18n(dataset, 'description', locale) ||
+        getOpenApiTextFromI18n(dataset, 'summary', locale) ||
+        getOpenApiTextFromI18n(dataset, 'description', locale) ||
         t('unnamed'),
       from: dataset.from,
     })),
@@ -604,8 +604,8 @@ export const ToolDialog = forwardRef<
                             }
                             label={
                               <Typography variant="caption">
-                                {getDatasetTextByI18n(parameter, 'description', locale) ||
-                                  getDatasetTextByI18n(parameter, 'name', locale)}
+                                {getOpenApiTextFromI18n(parameter, 'description', locale) ||
+                                  getOpenApiTextFromI18n(parameter, 'name', locale)}
                               </Typography>
                             }
                             labelPlacement="top"
@@ -621,8 +621,8 @@ export const ToolDialog = forwardRef<
             return (
               <Stack key={parameter.name}>
                 <Typography variant="caption" mb={0.5}>
-                  {getDatasetTextByI18n(parameter, 'description', locale) ||
-                    getDatasetTextByI18n(parameter, 'name', locale)}
+                  {getOpenApiTextFromI18n(parameter, 'description', locale) ||
+                    getOpenApiTextFromI18n(parameter, 'name', locale)}
                 </Typography>
 
                 <Controller
@@ -856,7 +856,6 @@ function AddSelectAgentPopperButton({
 
   const exists =
     assistant.type === 'router' ? new Set(Object.values(assistant.routes ?? {}).map((i) => i.data.id)) : new Set();
-  const groupByApi = groupBy(openApiOptions, (item) => (item?.id || '').split(':')?.[0] || '');
 
   if (!project) {
     return null;
@@ -908,35 +907,31 @@ function AddSelectAgentPopperButton({
           </List>
         </>
 
-        {Object.entries(groupByApi).map(([key, options]) => {
-          return (
-            <React.Fragment key={key}>
-              <GroupView name={key} description="Blocklet API">
-                <Box component={DiDAvatar} did={window.blocklet.appId} size={40} sx={{ borderRadius: 1 }} />
-              </GroupView>
+        <>
+          <GroupView name="Blocklet API" description="Blocklet API">
+            <Box component={DiDAvatar} did={window.blocklet.appId} size={40} sx={{ borderRadius: 1 }} />
+          </GroupView>
 
-              <List
-                dense
-                disablePadding
-                sx={{
-                  pl: 8,
-                  '>hr': { my: '0 !important', borderColor: 'grey.100', ml: 1 },
-                  '>hr:last-of-type': { display: 'none' },
-                }}>
-                {options.map((x) => {
-                  return (
-                    <MenuItem selected={exists.has(x.id)} key={x.id} onClick={() => onSelect?.(x)} sx={{ my: 0.25 }}>
-                      <Box flex={1}>{x.name || t('unnamed')}</Box>
-                      <Box sx={{ width: 40, textAlign: 'right' }}>
-                        {exists.has(x.id) && <Box component={Icon} icon={CheckIcon} />}
-                      </Box>
-                    </MenuItem>
-                  );
-                })}
-              </List>
-            </React.Fragment>
-          );
-        })}
+          <List
+            dense
+            disablePadding
+            sx={{
+              pl: 8,
+              '>hr': { my: '0 !important', borderColor: 'grey.100', ml: 1 },
+              '>hr:last-of-type': { display: 'none' },
+            }}>
+            {openApiOptions.map((x) => {
+              return (
+                <MenuItem selected={exists.has(x.id)} key={x.id} onClick={() => onSelect?.(x)} sx={{ my: 0.25 }}>
+                  <Box flex={1}>{x.name || t('unnamed')}</Box>
+                  <Box sx={{ width: 40, textAlign: 'right' }}>
+                    {exists.has(x.id) && <Box component={Icon} icon={CheckIcon} />}
+                  </Box>
+                </MenuItem>
+              );
+            })}
+          </List>
+        </>
 
         {!(agentOptions.length + openApiOptions.length) && (
           <>