+++ blocklets/ai-studio/vite.config.ts
@@ -1,9 +1,6 @@
 /* eslint-disable import/no-extraneous-dependencies */
-
 import { existsSync } from 'fs';
-import path from 'path';
 
-import buildOpenAPIPlugin from '@blocklet/dataset-sdk/plugin';
 import react from '@vitejs/plugin-react';
 import million from 'million/compiler';
 import { PluginOption, defineConfig } from 'vite';
@@ -42,7 +39,6 @@ export default defineConfig((config) => {
       config.command === 'build' && million.vite({ auto: true }),
       react(),
       createBlockletPlugin(),
-      buildOpenAPIPlugin({ apis: [path.join(__dirname, './api/src/routes/**/*.*')] }),
       svgr(),
     ],
     build: {