+++ packages/ai-runtime/src/core/executor/decision.ts
@@ -8,9 +8,7 @@ import {
   isChatCompletionChunk,
   isChatCompletionUsage,
 } from '@blocklet/ai-kit/api/types/index';
-import { getBuildInDatasets } from '@blocklet/dataset-sdk';
 import { getAllParameters, getRequiredFields } from '@blocklet/dataset-sdk/request/util';
-import { DatasetObject } from '@blocklet/dataset-sdk/types';
 import { call } from '@blocklet/sdk/lib/component';
 import { logger } from '@blocklet/sdk/lib/config';
 import { isNil } from 'lodash';
@@ -30,8 +28,7 @@ import { GetAgentResult, RunAssistantCallback, ToolCompletionDirective } from '.
 import { renderMessage } from '../utils/render-message';
 import { nextTaskId } from '../utils/task-id';
 import { toolCallsTransform } from '../utils/tool-calls-transform';
-import { AgentExecutorBase, AgentExecutorOptions } from './base';
-import { runAPITool } from './blocklet';
+import { AgentExecutorBase, AgentExecutorOptions, ExecutorContext } from './base';
 
 const md5 = (str: string) => crypto.createHash('md5').update(str).digest('hex');
 
@@ -46,14 +43,15 @@ export class DecisionAgentExecutor extends AgentExecutorBase {
 
     const message = await renderMessage(agent.prompt, inputs);
     const routes = agent?.routes || [];
-    const openApis = await getBuildInDatasets();
+
+    const blocklet = await this.context.getBlockletAgent(agent.id);
 
     logger.info('start get tool function');
     const toolAssistants = (
       await Promise.all(
         routes.map(async (tool) => {
           if (tool?.from === 'blockletAPI') {
-            const dataset = (openApis || []).find((x) => x.id === tool.id);
+            const dataset = (blocklet.openApis || []).find((x) => x.id === tool.id);
             if (!dataset) return undefined;
 
             const name = tool?.functionName || dataset.summary || dataset.description || '';
@@ -402,18 +400,21 @@ export class DecisionAgentExecutor extends AgentExecutorBase {
           };
 
           if (tool.tool.from === 'blockletAPI') {
-            return runAPITool({
-              tool: tool.tool,
-              taskId: currentTaskId,
-              assistant: agent,
-              parameters: { ...inputs, ...requestData },
-              dataset: tool.toolAssistant as DatasetObject,
-              parentTaskId: taskId,
-              callback: cb,
-              user: this.context.user,
-              sessionId: this.context.sessionId,
-              projectId: this.context.entryProjectId,
-            });
+            const blocklet = await this.context.getBlockletAgent(tool.tool.id);
+            if (!blocklet.agent) {
+              throw new Error('Blocklet agent api not found.');
+            }
+
+            const result = await this.context
+              .executor({ ...this.context, callback: cb } as ExecutorContext)
+              .execute(blocklet.agent, {
+                inputs: tool.tool.parameters,
+                variables: { ...inputs, ...requestData },
+                taskId: currentTaskId,
+                parentTaskId: taskId,
+              });
+
+            return result;
           }
 
           await Promise.all(