+++ packages/dataset-sdk/src/request/index.ts
@@ -1,4 +1,5 @@
 import { getComponentWebEndpoint } from '@blocklet/sdk/lib/component';
+import envConfig from '@blocklet/sdk/lib/config';
 import { sign } from '@blocklet/sdk/lib/util/verify-sign';
 import axios from 'axios';
 
@@ -9,30 +10,25 @@ export interface User {
   did: string;
 }
 
-export const getRequest = (
+export const callBlockletApi = (
   pathItem: DatasetObject,
-  requestData: { [key: string]: any },
+  data: { [key: string]: any },
   options?: { user?: User; params?: { [key: string]: any }; data?: { [key: string]: any } }
 ) => {
-  const requestConfig = getRequestConfig(pathItem, requestData, {
+  const requestConfig = getRequestConfig(pathItem, data, {
     params: options?.params || {},
     data: options?.data || {},
   });
   const { headers, body, ...config } = requestConfig;
-  let did = '';
-
-  if (options?.user) {
-    did = options?.user?.did || '';
-  }
-
-  config.baseURL = pathItem?.name ? getComponentWebEndpoint(pathItem?.name) : '';
+  const did = options?.user?.did || '';
 
   return axios({
     ...config,
+    baseURL: pathItem.name ? getComponentWebEndpoint(pathItem.name) : envConfig.env.appUrl,
     headers: {
       ...(headers || {}),
       'x-component-sig': sign(body || {}),
-      'x-component-did': process.env.BLOCKLET_COMPONENT_DID,
+      'x-component-did': pathItem.did || process.env.BLOCKLET_COMPONENT_DID,
       'x-user-did': did,
     },
   });