+++ packages/dataset-sdk/src/util/get-open-api-i18n-text.ts
@@ -1,0 +1,6 @@
+const getOpenApiTextFromI18n = (data: any, key: string, locale: string) => {
+  const i18nKey: string = `x-${key}-${locale}`;
+  return data[i18nKey] || data[key];
+};
+
+export default getOpenApiTextFromI18n;